Split into StreamInfo, ChatFeed, ChatInput
