+++ src/components/hooks/useOverrideBackPress.ts
@@ -1,0 +1,13 @@
+import { useCallback } from "react";
+import { BackHandler } from "react-native";
+import { useFocusEffect } from "@react-navigation/native";
+
+
+export default function useOverrideBackPress(onBackPress: () => boolean) {
+  useFocusEffect(
+    useCallback(() => {
+      const subscription = BackHandler.addEventListener('hardwareBackPress', onBackPress);
+      return () => subscription.remove();
+    }, [onBackPress])
+  );
+}
