+++ src/components/stream/ChatFeed.tsx
@@ -1,0 +1,10 @@
+import { View } from 'react-native';
+
+
+export default function ChatFeed() {
+  
+  return (
+    <View style={{flex: 1}}/>
+  );
+}
+
