+++ src/screens/StreamScreen.tsx
@@ -3,25 +3,23 @@ import {
   StyleSheet,
   View,
   Dimensions,
-  BackHandler,
-  TextInput,
-  TextInputSubmitEditingEvent,
   KeyboardAvoidingView,
   Keyboard,
-  Platform
 } from 'react-native';
-import { useFocusEffect } from '@react-navigation/native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
 import { StreamOverlayHandles, StreamScreenProps } from '../types';
 import { getStreamURL } from '../services/backend_service';
 import { Colors } from '../constants';
-import { showErrorSendingMessage, showErrorUnabletoStream, showSuccessSendingMessage } from '../alerts/alerts';
-import ChannelInfo from '../components/ChannelInfo';
-import { postMessage } from '../services/kick_service';
-import StreamPlayer from '../components/StreamPlayer';
-import StreamOverlay from '../components/StreamOverlay';
+import { showErrorUnabletoStream } from '../alerts/alerts';
+import StreamInfo from '../components/stream/StreamInfo';
+import StreamPlayer from '../components/stream/StreamPlayer';
+import StreamOverlay from '../components/stream/StreamOverlay';
+import ChatInput from '../components/stream/ChatInput';
+import ChatFeed from '../components/stream/ChatFeed';
+import useOverrideBackPress from '../components/hooks/useOverrideBackPress';
+import { GlobalKAVBehaviour } from '../helpers/helpers';
 
 
 const screenDimensions = Dimensions.get('screen');
@@ -31,7 +29,6 @@ export default function StreamScreen({ route }: StreamScreenProps) {
   const overlayRef = useRef<StreamOverlayHandles>(null);
 
   const [ streamURL, setStreamURL ] = useState<string>("");
-  const [ messageText, setMessageText ] = useState<string>("");
   const [ paused, setPaused ] = useState(false);
   const [ isFullscreen, setIsFullscreen ] = useState(false);
   const [ screenSize, setScreenSize ] = useState<{ width: number; height: number }>(screenDimensions);
@@ -57,26 +54,13 @@ export default function StreamScreen({ route }: StreamScreenProps) {
     };
   }, []);
 
-
-  useFocusEffect(
-    useCallback(() => {
-      const onBackPress = () => {
-        if (isFullscreen) {
-          overlayRef.current?.toggleFullscreen();
-          return true;
-        } else {
-          return false;
-        }
-      };
-
-      const subscription = BackHandler.addEventListener(
-        'hardwareBackPress',
-        onBackPress
-      );
-
-      return () => subscription.remove();
-    }, [isFullscreen])
-  );
+  useOverrideBackPress(useCallback(() => {
+    if (isFullscreen) {
+      overlayRef.current?.toggleFullscreen();
+      return true;
+    }
+    return false;
+  }, [isFullscreen]))
 
   const fetchStreamURL = () => {
     getStreamURL(channel.slug)
@@ -89,33 +73,13 @@ export default function StreamScreen({ route }: StreamScreenProps) {
   }
 
 
-
-
-  const handleSendMessage = (e: TextInputSubmitEditingEvent) => {
-    postMessage(
-      tokens.accessToken,
-      channel.id,
-      messageText.substring(0, 500),
-    ).then(() => {
-      // FIXME: Check response
-      showSuccessSendingMessage();
-    }).catch(() => {
-      showErrorSendingMessage();
-    });
-
-    setMessageText("");
-  }
-
-
   const videoWidth = screenSize.width;
   const videoHeight = isFullscreen ? screenSize.height : (videoWidth * 9) / 16;
-  const kavBehavior = Platform.OS === "ios" ? "padding" : "height";
   const WrapperView = isFullscreen ? View : SafeAreaView;
 
 
   return (
-    <WrapperView style={styles.container} >
-      <KeyboardAvoidingView style={{flex: 1}} behavior={kavBehavior} keyboardVerticalOffset={offset} >
+    <WrapperView style={styles.container}>
 
       <View style={[ styles.videoContainer, { width: videoWidth, height: videoHeight }]}>
       
@@ -124,7 +88,7 @@ export default function StreamScreen({ route }: StreamScreenProps) {
           paused={paused}
           setLoadingVideo={setLoadingVideo}
         />
-        
+
         <StreamOverlay
           ref={overlayRef}
           isStreamReady={isStreamReady}
@@ -134,25 +98,17 @@ export default function StreamScreen({ route }: StreamScreenProps) {
           setPaused={setPaused}
           setIsFullscreen={setIsFullscreen}
         />
- 
+
       </View>
 
       { !isFullscreen &&
-        <>
-          <ChannelInfo channel={channel} tokens={tokens} />
-          <View style={{ flex: 1, alignSelf: "stretch", justifyContent: "flex-end"}}>
-
-            <TextInput
-              style={styles.searchInput}
-              value={messageText}
-              onChangeText={setMessageText}
-              onSubmitEditing={handleSendMessage}
-            />
-
-          </View>
-        </>
+        <KeyboardAvoidingView style={{flex: 1}} behavior={GlobalKAVBehaviour} keyboardVerticalOffset={offset}>
+          <StreamInfo channel={channel} tokens={tokens} />
+          <ChatFeed/>
+          <ChatInput channel={channel} tokens={tokens} />
+        </KeyboardAvoidingView>
       }
-      </KeyboardAvoidingView>
+
     </WrapperView>
   );
 }
@@ -167,18 +123,5 @@ const styles = StyleSheet.create({
     backgroundColor: Colors.background,
     justifyContent: "center",
     alignItems: "center"
-  },
-  searchInput: {
-    height: 50,
-    backgroundColor: Colors.background,
-    borderColor: Colors.border,
-    color: Colors.textSecondary,
-    borderWidth: 1,
-    borderRadius: 5,
-    paddingHorizontal: 10,
-    marginHorizontal: 20,
-    marginBottom: 10,
-    fontSize: 16,
-    padding: 0,
   },
 });
